requirements.txt
absl-py==2.1.0
#accelerate==1.5.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
#click==8.1.8
datasets==3.3.2
dill==0.3.8
fastapi==0.115.11
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.71.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
#networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peft==0.14.0
propcache==0.3.0
#protobuf==6.30.0
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
#safetensors==0.5.3
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
#sympy==1.13.1
#tensorboard==2.19.0
#tensorboard-data-server==0.7.2
#tokenizers==0.21.0
torch==2.6.0
#tqdm
#transformers @ git+https://github.com/huggingface/transformers@46350f5eae87ac1d168ddfdc57a0b39b64b9a029
#triton==3.2.0
#trl==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
#google==3.0.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-aiplatform==1.93.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
#google-cloud-bigquery==3.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-scheduler
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
transformers
# PDF Processing
PyMuPDF==1.25.4
vllm
openai
